Release Version 23.03.2 (#11)
* update template for adding a database connection (#2)

* update template for adding a database connection

* prepare function to create a template for a new data source

* update function to generate templates, add tests, shift templates to test folder

* update description

* shift test helpers to R folder

* write templates to script, test this script

* split generateTemplateDataSource into two functions for file or db, update tests

* update file names

* update docu

* add set up of Renviron file, add tests

* add update of database list to both create*Source functions, and add tests and docu

* remove comment

* update addDBSettings, fix tests

* check for existing dbNames

* update the README with the new description for creating a new source

* fix style

* add note

* export functions

* format dbName for scripts (remove special characters)

* add logging of file creation or updates

* update description

* remove description param, update tests

* use template package to create template scripts

* template for file source

* rename dbName -> dataSourceName

* update Renviron for local connection to db

* template for db source

* template for Renviron

* remove old templates

* delete not needed example

* version up

* update readme

* fix test after auto format

* fix branch name in readme

* change order of sections

* update comments in templates and readme

* Update README.md

* Update R/01-createNewFileSource.R


* Update R/01-createNewFileSource.R


* sheetNumber instead of sheetName

* update readme

* update readme

* update filepaths and fix tests

* create a news.md

* missing sendReport function (#8)

* missing sendReport function

* add mappingId

* Update R/01-etl.R


* Update R/01-etl.R


* Update R/01-etl.R


---------


* update branches in Deployment section

* Feat/4 add new mappings (#6)

* parameterize getMappingTable

* test mapping (in Lives removed site, dateLower, dateUpper)

* add some notes what is next

* add new parameter mappingName to templates

* add parameter mappingName to list of databases

* add parameter mapping to more functions

* add parameter sources to etlMapping()

* remove source specific load functions

* make defaultVars depend on mapping

* update tests

* update tests for createNewFileSource

* remove global variable

* docu for prepareData

* add .data

* fix docu

* fix db warning

* new function mappingNames()

* still clean old tables

* notes

* clean up code

* fix test

* reorder helpers

* update docu

* extract fct setDataSourceName, restructure inheritParams, update test

* restructure inheritParams

* add comments

* fix toUpper of dataSourceName

* extract checkDataSourceName

* auto format

* add load_all

* fix toUpper for dataSourceNames and tests

* new param isTest

* add more logging

* rename files

* prepare updateOrCreateTableMPI function

* shift updateOrCreateTbale, fix test

* rename templates

* create only table data if not exists; update all others

* fix parameters

* update cleanUp

* remove db from mapping query

* replace temporary name for the mapping where possible

* add function mappingNames to tests

* update databases function

* functionality to create a new table if not exists

* update readMe

* update news.md

* remove duplicated

* Update README.md

* debug with some logging

* shift queries into sql folder, fix sql

* fix global var error

* fix wrong package name

* fix queries

* add logging

* update checkout, login-action, metadata-action & build-push-action to newer version & enable running workflow manually

* update checkout, login-action, metadata-action & build-push-action to newer version & enable running workflow manually

* switch to posit cran mirror

* undo mran / posit mirror fix

* fix typo in sql queries (#10)

* fix ticks, add sendReport, exclude columns extraVars

* update test

* Keep updating the deprecated table 'data'

* fix errors

* version up

* update logging

* fix typo in message for latitude

* fix case if is.na(coordType)

* try new test

* remove test

---------

+++ R/03-clean.R
@@ -1,20 +1,28 @@
 #' Clean up database
 #'
+#' @param mappings mappings to be used
+#' @param sources sources to be used
+#'
 #' @export
-cleanUp <- function(){
-  sendQueryMPI(cleanUpQry("data"))
-  # update also new table 'IsoMemo_data'
-  sendQueryMPI(cleanUpQry("IsoMemo_data"))
-  sendQueryMPI(cleanUpQry("extraCharacter"))
-  sendQueryMPI(cleanUpQry("extraNumeric"))
-  sendQueryMPI(cleanUpQry("warning"))
-}
-
-cleanUpQry <- function(table, sources = dbnames()){
+cleanUp <- function(mappings = mappingNames(), sources = dbnames()){
+  # delete data from removed sources
   sources <- dbtools::sqlParan(sources, function(x) dbtools::sqlEsc(x, with = "'"))
-  dbtools::Query(
-    "DELETE FROM {{ table }} where `source` not in {{ sources }};",
-    table  = table,
-    sources = sources
-  )
+
+  for (mappingName in mappings) {
+    if (mappingName == "IsoMemo") {
+      # Keep cleaning the deprecated table 'data' until the CRAN package,
+      #  the API and the app are updated to use only the new table 'IsoMemo_data'!
+      dbtools::Query(
+        "DELETE FROM {{ table }} where `source` not in {{ sources }};",
+        table = "data",
+        sources = sources
+      ) %>%
+        sendQueryMPI()
+    }
+
+    sendQueryMPI("removeOldSourceFromData", mappingId  = mappingName, sources = sources)
+    sendQueryMPI("removeOldSourceFromTable", mappingId = mappingName, table = "extraCharacter", sources = sources)
+    sendQueryMPI("removeOldSourceFromTable", mappingId = mappingName, table = "extraNumeric", sources = sources)
+    sendQueryMPI("removeOldSourceFromTable", mappingId = mappingName, table = "warning", sources = sources)
+  }
 }
